feat(playground): 实现模板服务并重构编辑器状态管理
- 新增 TemplateService 类，用于统一管理代码模板
- 重构 EditorStore，使用模板服务替代原有的示例代码
- 优化编辑器初始化状态，使用模板代码作为初始内容
- 更新切换语言功能，改为使用模板代码替换示例代码
