+++ apps/playground/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "playground-next",
+  "name": "playground",
   "version": "0.1.0",
   "private": true,
   "scripts": {
